+++ src/shared/components/Configuracoes/ChangePassword.tsx
@@ -1,0 +1,115 @@
+import { Box, Button, FormControl, IconButton, InputAdornment, InputLabel, OutlinedInput, Paper, Typography } from "@mui/material";
+import { useState } from "react";
+import VisibilityIcon from '@mui/icons-material/Visibility';
+import VisibilityOffIcon from '@mui/icons-material/VisibilityOff';
+import SaveIcon from '@mui/icons-material/Save';
+
+interface ITrocaSenha {
+    password: string
+    showPassword: boolean;
+    cPassword: string;
+    showCPassword: boolean;
+}
+
+const ChangePassword: React.FC = () => {
+    const [valuesTC, setValuesTC] = useState<ITrocaSenha>({
+        password: '',
+        showPassword: false,
+        cPassword: '',
+        showCPassword: false
+
+    });
+
+
+    const handleChangeTC = (prop: keyof ITrocaSenha) => (event: React.ChangeEvent<HTMLInputElement>) => {
+        setValuesTC({ ...valuesTC, [prop]: event.target.value });
+    };
+
+    const handleClickShowPassword = () => {
+        setValuesTC({
+            ...valuesTC,
+            showPassword: !valuesTC.showPassword,
+        });
+    };
+
+    const handleClickShowTCCPassword = () => {
+        setValuesTC({
+            ...valuesTC,
+            showCPassword: !valuesTC.showCPassword,
+        });
+    };
+
+    const handleMouseDownPassword = (event: React.MouseEvent<HTMLButtonElement>) => {
+        event.preventDefault();
+    };
+
+    const handleChangePassword = () => {
+        alert('Implementar troca de senha!');
+    }
+
+    return (
+        <Box
+            gap={1}
+            margin={1}
+            padding={1}
+            paddingX={2}
+            height="auto"
+            display="flex"
+            flexDirection="column"
+            alignItems="start"
+            component={Paper} >
+            <Typography variant='h6'>
+                Trocar Senha
+            </Typography>
+            <FormControl size="small" fullWidth variant="outlined" >
+                <InputLabel htmlFor="txtTCPassword">Senha</InputLabel>
+                <OutlinedInput
+                    id="txtTCPassword"
+                    type={valuesTC.showPassword ? 'text' : 'password'}
+                    value={valuesTC.password}
+                    onChange={handleChangeTC('password')}
+                    endAdornment={
+                        <InputAdornment position="end">
+                            <IconButton
+                                aria-label="toggle password visibility"
+                                onClick={handleClickShowPassword}
+                                onMouseDown={handleMouseDownPassword}
+                                edge="end"
+                            >
+                                {valuesTC.showPassword ? <VisibilityOffIcon /> : <VisibilityIcon />}
+                            </IconButton>
+                        </InputAdornment>
+                    }
+                    label="Senha"
+                    inputProps={{ maxLength: 50 }}
+                />
+            </FormControl>
+            <FormControl size="small" fullWidth variant="outlined" >
+                <InputLabel htmlFor="txtConfirmPassword">Confirma Senha</InputLabel>
+                <OutlinedInput
+                    id="txtConfirmPassword"
+                    type={valuesTC.showCPassword ? 'text' : 'password'}
+                    value={valuesTC.cPassword}
+                    onChange={handleChangeTC('cPassword')}
+                    endAdornment={
+                        <InputAdornment position="end">
+                            <IconButton
+                                aria-label="toggle password visibility"
+                                onClick={handleClickShowTCCPassword}
+                                onMouseDown={handleMouseDownPassword}
+                                edge="end" >
+                                {valuesTC.showPassword ? <VisibilityOffIcon /> : <VisibilityIcon />}
+                            </IconButton>
+                        </InputAdornment>
+                    }
+                    label="Confirma Senha"
+                    inputProps={{ maxLength: 50 }}
+                />
+            </FormControl>
+            <Button color='primary' disableElevation variant='contained' startIcon={<SaveIcon />} onClick={handleChangePassword} >Salvar</Button>
+        </Box>
+
+
+    );
+}
+export default ChangePassword;
